Estoque: fecha a janela de corrupcao aberta na v5.58 (validacao relaxada para numeric sem o alargamento aplicado, com o MySQL arredondando 0,5 para 1 em silencio) e elimina 10 truncamentos de quantidade remanescentes, incluindo o PDF de orcamento que o cliente assina. Migration emendada para preservar nulidade e defaults das colunas legadas em vez de impor NOT NULL DEFAULT 0
v5.58.1.0 (patch) — commit gerado por scripts/bash/deploy-completo.sh

+++ backend/app/Services/Budgets/BudgetWorkflowService.php
@@ -371,7 +371,7 @@ class BudgetWorkflowService
                 'categoria' => (string) ($peca->categoria ?? ''),
                 'preco_custo' => (float) ($peca->preco_custo ?? 0),
                 'preco_venda' => (float) ($peca->preco_venda ?? 0),
-                'quantidade_atual' => (int) ($peca->quantidade_atual ?? 0),
+                'quantidade_atual' => (float) ($peca->quantidade_atual ?? 0),
             ])
             ->values()
             ->all();
